Cambia de directorio todos los containers a features con todos sus componentes no reutilizables
