+++ src/app/(app)/prp/(main-layout)/expensas/[id]/reclamos/[rclid]/page.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import NoResult from '@/containers/errors/no-result';
-import RclDetails from '@/containers/rcl-page/rcl-details'
+import RclDetails from 'app/features/Issues/Details'
 import getParams from '@/env/getParams';
 import { getUnitIssues } from '@/lib/queries/queries';
+import EmptySection from '@/components/Sections/AppSections/Errors/EmptySection';
 
 const page = async ({ params: { rclid, id } }: { params: { rclid: string, id: string } }) => {
   const unitCode = getParams({ params: id, type: "code" });
@@ -12,7 +12,7 @@ const page = async ({ params: { rclid, id } }: { params: { rclid: string, id: st
   const data = await getUnitIssues({ code: unitCode, id: rclid });
 
   if (data.length === 0) {
-    return <NoResult message={'No se han encontrado resultados'}/>
+    return <EmptySection title={'No se han encontrado resultados'}/>
   }
 
   const details = data && (data.rclDpto.length === 0 ? data.rclEdif : data.rclDpto)
