+++ src/app/(app)/prp/(main-layout)/expensas/[id]/reclamos/nuevo/page.tsx
@@ -1,4 +1,4 @@
-import NewRcl from '@/containers/rcl-page/new-rcl';
+import NewRcl from 'app/features/Issues/Create';
 
 const page = ({ params }: { params: { id: string } }) => {
   return <NewRcl id={params.id} />;
