+++ src/app/(app)/prp/(main-layout)/expensas/page.tsx
@@ -1,10 +1,8 @@
 import React from 'react'
-import { getUnits } from '@/lib/queries/queries';
 import Section from '@/components/Sections/AppSections/Section';
-import Grid from '@/components/Sections/AppSections/Grid';
-import UnitCard from '@/components/Cards/UnitCard';
-import NewUnitDialog from '@/components/Dialogs/NewUnit';
+import { getUnits } from '@/lib/queries/queries';
 import { redirect } from 'next/navigation';
+import UnitList from 'app/features/Unit/List';
 
 const page = async () => {
   const data = await getUnits();
@@ -15,18 +13,7 @@ const page = async () => {
 
   return (
     <Section className="w-full col-span-2 max-md:col-span-1 pb-8 max-md:pb-0">
-      <Grid className="max-md:pb-6">
-        {data.map((unit: any) => {
-          return <UnitCard unit={unit} key={unit.uf_id} />;
-        })}
-        <div
-          className={
-            "border border-outline dark:border-outline-dark border-dashed rounded-lg flex items-center justify-center max-lg:border-0 py-28 max-lg:py-0"
-          }
-        >
-          <NewUnitDialog />
-        </div>
-      </Grid>
+      <UnitList data={data}/>
     </Section>
   );
 }
