--- src/containers/expenses-page/aside-section/index.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const AsideSection = () => {
-  return (
-    <>
-      
-    </>
-  )
-}
-
-export default AsideSection
