--- src/containers/modals/shortcuts-actions/index.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const index = () => {
-  return (
-    <div>index</div>
-  )
-}
-
-export default index
