--- src/containers/news-page/news-list/index.tsx
@@ -1,19 +1,0 @@
-import React from 'react'
-import NewCard from '@/components/Cards/NewCard';
-
-interface NewsScreenProps {
-  data: any[];
-  param?: string;
-}
-
-const NewsScreen = ({ data, param }: NewsScreenProps) => {
-  return (
-    <div className="flex flex-col gap-[10px]">
-      {data.map((item, index) => {
-        return <NewCard item={item} index={index} key={`new_${item.id}`} />;
-      })}
-    </div>
-  );
-}
-
-export default NewsScreen
